requirements_prod.txt
Django==5.0.8
PyMySQL==1.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
